Move onboarding base class to subfolder
