+++ cypress/support/pages/onboarding/onboarding_page.js
